Update adding-staging-commiting.md
Yulric: when I click on the gitter.md file link in this document, it leads to an error page. But if I go back to the list of files and click on gitters.md, then it works. So perhaps it's just the file path link in this document that doesn't work.